requirements.txt
app==0.0.1
awscli==1.16.174
boto3==1.9.164
botocore==1.12.164
certifi==2019.3.9
Click==7.0
colorama==0.3.9
cycler==0.10.0
DateTime==4.3
docutils==0.14
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
MarkupSafe==1.1.1
numpy==1.16.4
pandas==0.24.2
pyasn1==0.4.5
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.4
urllib3==1.25.3
virtualenv==16.6.0
Werkzeug==0.15.4
zope.interface==4.6.0
